+++ src/enums/constants.ts
@@ -1,0 +1,6 @@
+export enum FileSize {
+    TENTH_MB = 1024 * 103,
+    HALF_MB = 1024 * 512,
+    ONE_MB = 1024 * 1024 * 1,
+    TEN_MB = 1024 * 1024 * 10,
+}